Adding Code Events
lots of stuff to do for a Code
After ACLS training, though why no that patients that have a code
situation and start a organized code game

actions -> cause changes in the state machine that is the patient.

Needs to complete all the actions in the right order / time in order to
save th patient.